+++ tests/src/handlers/test_functions.py
@@ -73,7 +73,7 @@ class TestRequirePasswordDecorator:
 
 class TestSearchHandler:
     def test_kp_client_error(self, mocker, valid_settings, keepassxc_client):
-        mocker.patch.object(keepassxc_client, "locate", side_effect=OSError)
+        mocker.patch.object(keepassxc_client, "search", side_effect=OSError)
         mocker.patch("handlers.initialize_keepassxc_client", return_value=keepassxc_client)
         add_item_mock = mocker.patch("handlers.AlfredScriptFilter.add_item")
         send_mock = mocker.patch("handlers.AlfredScriptFilter.send")
@@ -100,7 +100,7 @@ class TestSearchHandler:
         self, mocker, environ_factory, valid_settings, keepassxc_client, kp_output, expected_title, expected_arg
     ):
         environ_factory(entry_delimiter=" > ")
-        mocker.patch.object(keepassxc_client, "locate", return_value=[kp_output])
+        mocker.patch.object(keepassxc_client, "search", return_value=[kp_output])
         mocker.patch("handlers.initialize_keepassxc_client", return_value=keepassxc_client)
         add_item_mock = mocker.patch("handlers.AlfredScriptFilter.add_item")
         send_mock = mocker.patch("handlers.AlfredScriptFilter.send")